+++ src/index.js
@@ -1,8 +1,8 @@
-export { default as Form } from './form'
-export { default as Field } from './field'
-export { default as FieldType } from './field-type'
-export { default as ArrayComponent } from './array'
-export { default as ObjectComponent } from './object'
+export {default as Form} from './form'
+export {default as Field} from './field'
+export {default as FieldType} from './field-type'
+export {default as ArrayComponent} from './array'
+export {default as ObjectComponent} from './object'
 export {
   registerType,
   getFieldType,